requirements.txt
Flask
Flask-Babel
pdf2image
python-pptx
Pillow
Werkzeug